[BRANCH START] Moved Create_funcs
